Integrar autenticacion de usuarios requerida
Requerir que un usuario esté autenticado y con sesion activa
Para poder acceder a las funcionalidades de crear, ver y editar reportes

+++ src/routes/reportes.py
@@ -1,21 +1,34 @@
 from flask import (
     Blueprint, flash, render_template,
-    request, redirect, url_for
+    request, redirect, url_for, abort
 )
+from flask_login import current_user, login_required
 from src.models.reporte import Reporte
 from src.models.comentario import Comentario
 from src.forms.reportes import CrearReporte
 from src import db
+from sqlalchemy.sql import func
 
 reportes = Blueprint('reportes', __name__, url_prefix='/reportes',
                     template_folder='templates')
 
+@reportes.route('/ver-reportes', methods=['GET'])
+@login_required
+def ver_reportes():
+    print(current_user, current_user.id)
+
+    reportes = Reporte.query.all()
+    
+    return render_template('reportes/ver-reportes.html', data={'reportes': reportes})
+
 # from sqlalchemy.orm import asdict
 @reportes.route('/reporte/<string:id>')
+@login_required
 def ver_reporte(id):
     # crear_comentario.reporte_id = id
     reporte = Reporte.query.get_or_404(id)
     comentarios = Comentario.query.filter_by(reporte_id=id)#.first()
+    # reporte.fecha_atencion = func.now()
     
     forms = {'crear_comentario': crear_comentario}
     return render_template('reportes/ver-reporte.html',
@@ -23,8 +36,8 @@ def ver_reporte(id):
                            forms=forms)
     
 
-
 @reportes.route('/crear-reporte', methods=['GET', 'POST'])
+@login_required
 def crear_reporte():
     form = CrearReporte()
 
@@ -34,7 +47,8 @@ def crear_reporte():
         tipo = request.form['tipo']
         categoria = request.form['categoria']
         
-        reporte = Reporte(titulo=titulo, descripcion=descripcion, tipo=tipo, categoria=categoria)
+        reporte = Reporte(titulo=titulo, descripcion=descripcion, tipo=tipo, categoria=categoria,
+                          usuario_id=current_user.id)
         db.session.add(reporte)
         db.session.commit()
         
@@ -49,8 +63,11 @@ def crear_reporte():
 
 
 @reportes.route('/editar-reporte/<string:id>', methods=['GET', 'POST'])
+@login_required
 def editar_reporte(id):
     reporte = Reporte.query.get_or_404(id)
+    if reporte.id != current_user.id:
+        abort(403)
     if request.method == 'POST':
         titulo = request.form['titulo']
         descripcion = request.form['descripcion']
@@ -72,9 +89,13 @@ def editar_reporte(id):
     
     
 @reportes.route('/eliminar-reporte/<string:id>', methods=['GET', 'POST'])
+@login_required
 def eliminar_reporte(id):
     if request.method == 'POST':
         reporte = Reporte.query.get_or_404(id)
+        if reporte.id != current_user.id:
+            abort(403)
+
         db.session.delete(reporte)
         db.session.commit()
         
@@ -88,23 +109,17 @@ def eliminar_reporte(id):
 
         return redirect(url_for('reportes.ver_reportes'))
         
-        
-    
-@reportes.route('/ver-reportes', methods=['GET'])
-def ver_reportes():
-    reportes = Reporte.query.all()
-    
-    return render_template('reportes/ver-reportes.html', data={'reportes': reportes})
-
 
 # @reportes.route("/crear-comentario/<int:reporte_id>", methods=["POST"])
 @reportes.route('/reporte/<int:reporte_id>/crear-comentario', methods=["POST"])
+@login_required
 def crear_comentario(reporte_id):
     if request.method == 'POST':
         comentario_cont = request.form['comentario']
         # reporte_id = request.form['reporte_id']
         
-        comentario = Comentario(comentario=comentario_cont, reporte_id=reporte_id)
+        comentario = Comentario(comentario=comentario_cont, reporte_id=reporte_id,
+                                usuario_id=current_user.id)
 
         db.session.add(comentario)
         db.session.commit()
@@ -114,9 +129,13 @@ def crear_comentario(reporte_id):
     return redirect(url_for("reportes.ver_reporte", id=reporte_id))
 
 @reportes.route('/reporte/<int:reporte_id>/eliminar-comentario/<int:id>', methods=["POST"])
+@login_required
 def eliminar_comentario(id, reporte_id):
     if request.method == 'POST':
         comentario = Comentario.query.get_or_404(id)
+        if comentario.id != current_user.id:
+            abort(403)
+            
         db.session.delete(comentario)
         db.session.commit()
